+++ ostap/fitting/roofitresult.py
@@ -650,8 +650,8 @@ def _rfr_table_ ( r , title = '' , prefix = '' , more_vars = {} ) :
     if 0 < nbadnll :
         rows.append ( ( 'Invalid FCN/NLL evaluations' , '' , '  %d' % nbadnll , '' ) )
 
-    ## rows = [ ( '', 'Unit', 'Value' , 'Global/max correlation [%]') ] + rows
-    rows = [ ( '', 'Unit', 'Value' , 'Max correlation [%]') ] + rows
+    if (6,26) <= root_info : rows = [ ( '', 'Unit', 'Value' , 'Global/max correlation [%]') ] + rows
+    else                   : rows = [ ( '', 'Unit', 'Value' , 'Max correlation [%]') ] + rows
 
     pars_all   = r.params ( float_only = False )
     pars_float = r.params ( float_only = True  )
@@ -690,16 +690,18 @@ def _rfr_table_ ( r , title = '' , prefix = '' , more_vars = {} ) :
         if 0 <= cq and not  ( 6 , 24 ) <= root_info < ( 6 , 25 ) and 1 < len ( pars_float ) :
             
             mxr , mxv = r.max_cor    ( p )
-            
-            ## logger.warning ("globalCorr.... %s %s" %  ( p , cq ) )
-            gc = -1.0 
-            ## gc    = r.globalCorr ( p ) if 3 == cq else -1.00
-            ## cc        = '% +5.1f/(% +5.1f,%s)' % ( gc*100 , mxr*100 , mxv )
 
-            cc        = '% +5.1f : %-s' % ( mxr*100 , mxv )
-            if 0.95 < abs ( mxr ) : cc = attention ( cc )
+            if (6,26)<= root_info : 
+                gc = -1.0 
+                gc    = r.globalCorr ( p ) if 3 == cq else -1.00
+                if 0 <= gc :  cc = '% +5.1f/(% +5.1f,%s)' % ( gc*100 , mxr*100   , mxv )
+                else       :  cc = '% +5.1f : %-s'        % (          mxr * 100 , mxv )                
+                if 0.95 < abs ( mxr ) or 0.95 < gc : cc = attention ( cc )
+                
+            else : 
+                cc        = '% +5.1f : %-s' % ( mxr * 100 , mxv )                
+                if 0.95 < abs ( mxr ) : cc = attention ( cc )
             
-            ## if 0.95 < abs ( gc ) or 0.95 < abs ( mxr ) : cc = attention ( cc )
             max_corr = True
 
         row = p , n , s , cc