+++ ReleaseNotes/release_notes.md
@@ -1,4 +1,6 @@
 ## New features: 
+   1. add resolution model based on Generalised Gaussian v1 distribution
+   1. re-enable pritout of the Global correlation for `RooFitResult` objects 
 
 ## Backward incompatible: